% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_exposure_rate}
\alias{add_exposure_rate}
\title{add_exposure_rate}
\usage{
add_exposure_rate(.data, fit)
}
\arguments{
\item{.data}{A data frame}

\item{fit}{A fit object}
}
\value{
A \code{tbl}
}
\description{
add_exposure_rate
}
\keyword{internal}
