% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_partition}
\alias{add_partition}
\title{add_partition}
\usage{
add_partition(df.input, partition_by, n_partitions)
}
\arguments{
\item{df.input}{A tibble}

\item{partition_by}{A symbol. Column we want to partition by}

\item{n_partitions}{An integer number of partition}
}
\description{
Add partition column dto data frame
}
