% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{check_if_any_NA}
\alias{check_if_any_NA}
\title{Check if NA}
\usage{
check_if_any_NA(input.df, sample_column, gene_column, value_column,
  significance_column, do_check_column, formula_columns)
}
\description{
Check if NA
}
