% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts.ppc}
\alias{counts.ppc}
\title{counts.ppc}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
counts.ppc
}
\description{
Contains an example result for the vignette
}
\keyword{datasets}
