% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts}
\alias{counts}
\title{Example data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 394821 rows and 9 columns.
}
\usage{
counts
}
\description{
Example data set
}
\keyword{datasets}
