% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Create the design matrix}
\usage{
create_design_matrix(input.df, formula, sample_column)
}
\arguments{
\item{input.df}{A tibble}

\item{formula}{A formula}

\item{sample_column}{A symbol}
}
\description{
Create the design matrix
}
