% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_optimal_number_of_chains}
\alias{find_optimal_number_of_chains}
\title{Choose the number of chains baed on how many draws we need from the posterior distribution
Because there is a fix cost (warmup) to starting a new chain,
we need to use the minimum amount that we can parallelise}
\usage{
find_optimal_number_of_chains(
  how_many_posterior_draws,
  max_number_to_check = 100
)
}
\arguments{
\item{how_many_posterior_draws}{A real number of posterior draws needed}

\item{max_number_to_check}{A sane upper plateau}
}
\value{
A Stan fit object
}
\description{
Choose the number of chains baed on how many draws we need from the posterior distribution
Because there is a fix cost (warmup) to starting a new chain,
we need to use the minimum amount that we can parallelise
}
