% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{format_for_MPI}
\alias{format_for_MPI}
\title{format_for_MPI}
\usage{
format_for_MPI(df, shards, sample_column)
}
\arguments{
\item{df}{A tibble}

\item{shards}{A integer}

\item{sample_column}{A symbol}
}
\description{
Format reference data frame for MPI
}
