% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_input}
\alias{format_input}
\title{Format the input}
\usage{
format_input(
  .data,
  formula,
  .sample,
  .transcript,
  .abundance,
  .do_check,
  .significance,
  how_many_negative_controls = 500
)
}
\arguments{
\item{.data}{A tibble including a gene name column | sample name column | read counts column | covariates column}

\item{formula}{A formula}

\item{.sample}{A column name}

\item{.transcript}{A column name}

\item{.abundance}{A column name}

\item{.do_check}{A symbol}

\item{.significance}{A column name}

\item{how_many_negative_controls}{An integer}
}
\description{
Format the input
}
