% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{get_outlier_data_to_exlude}
\alias{get_outlier_data_to_exlude}
\title{Identify the optimal number of chain
based on how many draws we need from the posterior}
\usage{
get_outlier_data_to_exlude(counts_MPI, to_exclude, shards)
}
\arguments{
\item{counts_MPI}{A matrix of read count information}

\item{to_exclude}{A vector of oulier data points to exclude}

\item{shards}{An integer}
}
\value{
A matrix
}
\description{
Identify the optimal number of chain
based on how many draws we need from the posterior
}
