% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_approx_bias_lower}
\alias{lm_approx_bias_lower}
\title{Linear model lower}
\format{An object of class \code{lm} of length 12.}
\usage{
lm_approx_bias_lower
}
\description{
Linear model lower
}
\keyword{datasets}
