% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_approx_bias_upper}
\alias{lm_approx_bias_upper}
\title{Linear model upper}
\format{An object of class \code{lm} of length 12.}
\usage{
lm_approx_bias_upper
}
\description{
Linear model upper
}
\keyword{datasets}
