% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{ppc_seq}
\alias{ppc_seq}
\title{pcc_seq main}
\usage{
ppc_seq(input.df, formula = ~1, sample_column = "sample",
  gene_column = "symbol", value_column = "read count",
  significance_column = "p-value", do_check_column, full_bayes = F)
}
\arguments{
\item{input.df}{A tibble including a gene name column | sample name column | read counts column | covariates column}

\item{formula}{A formula}

\item{sample_column}{A character string}

\item{gene_column}{A character string}

\item{value_column}{A character string}

\item{full_bayes}{A boolean}
}
\value{
A tibble with additional columns
}
\description{
This function calls the stan model.
}
