% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq-package.R
\docType{package}
\name{ppcSeq-package}
\alias{ppcSeq-package}
\alias{ppcSeq}
\title{The 'ppcSeq' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
