% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{produce_plots}
\alias{produce_plots}
\title{Produce generated quantities plots with marked uotliers}
\usage{
produce_plots(.x, symbol, .abundance, .sample, covariate)
}
\arguments{
\item{.x}{A tibble}

\item{symbol}{A symbol object}

\item{.abundance}{A symbol object}

\item{.sample}{A symbol object}

\item{covariate}{A character string}
}
\value{
A ggplot
}
\description{
Produce generated quantities plots with marked uotliers
}
