% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{select_to_check_and_house_keeping}
\alias{select_to_check_and_house_keeping}
\title{Select only significant genes plus background for efficient normalisation}
\usage{
select_to_check_and_house_keeping(
  .data,
  .do_check,
  .significance,
  .transcript,
  how_many_negative_controls = 500
)
}
\arguments{
\item{.data}{A tibble}

\item{.do_check}{A boolean}

\item{.significance}{A symbol}

\item{.transcript}{A column name}

\item{how_many_negative_controls}{An integer}
}
\value{
A \code{tbl}
}
\description{
Select only significant genes plus background for efficient normalisation
}
\keyword{internal}
