% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{vb_iterative}
\alias{vb_iterative}
\title{vb_iterative}
\usage{
vb_iterative(model, output_samples, iter, tol_rel_obj)
}
\arguments{
\item{model}{A Stan model}

\item{output_samples}{An integer of how many samples from posteriors}

\item{iter}{An integer of how many max iterations}

\item{tol_rel_obj}{A real}
}
\value{
A Stan fit object
}
\description{
Runs iteratively variational bayes until it suceeds
}
