% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_attr}
\alias{add_attr}
\title{Add attribute}
\usage{
add_attr(var, attribute, name)
}
\arguments{
\item{var}{A character}

\item{attribute}{An object}

\item{name}{A character}
}
\value{
Same object
}
\description{
Add attribute
}
\keyword{internal}
