% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Create the design matrix}
\usage{
create_design_matrix(.data, formula, .sample)
}
\arguments{
\item{.data}{A tibble}

\item{formula}{A formula}

\item{.sample}{A symbol}
}
\value{
A \code{matrix}
}
\description{
Create the design matrix
}
\keyword{internal}
