% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{do_inference}
\alias{do_inference}
\title{do_inference}
\usage{
do_inference(
  my_df,
  formula,
  .sample,
  .transcript,
  .abundance,
  .significance,
  .do_check,
  approximate_posterior_inference = F,
  approximate_posterior_analysis = F,
  do_correct_approx = T,
  C,
  X,
  lambda_mu_mu,
  cores,
  exposure_rate_multiplier,
  intercept_shift_scale,
  additional_parameters_to_save,
  adj_prob_theshold,
  how_many_posterior_draws,
  to_exclude = tibble(S = integer(), G = integer()),
  truncation_compensation = 1,
  save_generated_quantities = F,
  inits_fx = "random",
  prior_from_discovery = tibble(.variable = character(), mean = numeric(), sd =
    numeric()),
  pass_fit = F,
  tol_rel_obj = 0.01,
  write_on_disk = F,
  seed
)
}
\arguments{
\item{my_df}{A tibble including a gene name column | sample name column | read counts column | covariates column}

\item{formula}{A formula}

\item{.sample}{A column name}

\item{.transcript}{A column name}

\item{.abundance}{A column name}

\item{.significance}{A column name}

\item{.do_check}{A column name}

\item{approximate_posterior_inference}{A boolean}

\item{approximate_posterior_analysis}{A boolean}

\item{do_correct_approx}{A boolean}

\item{C}{An integer}

\item{X}{A tibble}

\item{lambda_mu_mu}{A real}

\item{cores}{An integer}

\item{exposure_rate_multiplier}{A real}

\item{intercept_shift_scale}{A real}

\item{additional_parameters_to_save}{A character vector}

\item{adj_prob_theshold}{A real}

\item{how_many_posterior_draws}{A real number of posterior draws needed}

\item{to_exclude}{A boolean}

\item{truncation_compensation}{A real}

\item{save_generated_quantities}{A boolean}

\item{inits_fx}{A function}

\item{prior_from_discovery}{A tibble}

\item{pass_fit}{A fit}

\item{tol_rel_obj}{A real}

\item{write_on_disk}{A boolean}

\item{seed}{an integer}
}
\value{
A tibble with additional columns
}
\description{
This function calls the stan model.
}
