% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{fit_to_counts_rng_approximated}
\alias{fit_to_counts_rng_approximated}
\title{fit_to_counts_rng_approximated}
\usage{
fit_to_counts_rng_approximated(
  fit,
  adj_prob_theshold,
  how_many_posterior_draws,
  truncation_compensation,
  do_correct_approx = T,
  cores
)
}
\arguments{
\item{fit}{A fit object}

\item{adj_prob_theshold}{fit real}

\item{how_many_posterior_draws}{An integer}

\item{do_correct_approx}{A boolean}

\item{cores}{An integer}
}
\description{
fit_to_counts_rng_approximated
}
