% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_for_MPI}
\alias{format_for_MPI}
\title{format_for_MPI}
\usage{
format_for_MPI(df, shards, .sample)
}
\arguments{
\item{df}{A tibble}

\item{shards}{A integer}

\item{.sample}{A symbol}
}
\value{
A \code{tbl}
}
\description{
Format reference data frame for MPI
}
\keyword{internal}
