% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{format_input}
\alias{format_input}
\title{Format the input}
\usage{
format_input(input.df, formula, sample_column, gene_column, value_column,
  do_check_column, significance_column, how_many_negative_controls)
}
\arguments{
\item{input.df}{A tibble including a gene name column | sample name column | read counts column | covariates column}

\item{formula}{A formula}

\item{sample_column}{A column name}

\item{gene_column}{A column name}

\item{value_column}{A column name}

\item{do_check_column}{A symbol}

\item{significance_column}{A column name}

\item{how_many_negative_controls}{An integer}
}
\description{
Format the input
}
