% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R, R/ppcseq.R
\name{identify_outliers}
\alias{identify_outliers}
\title{pcc_seq main}
\usage{
identify_outliers(
  .data,
  formula = ~1,
  .sample,
  .transcript,
  .abundance,
  .significance,
  .do_check,
  percent_false_positive_genes = "1\%",
  how_many_negative_controls = 500,
  approximate_posterior_inference = T,
  approximate_posterior_analysis = NULL,
  draws_after_tail = 10,
  save_generated_quantities = F,
  additional_parameters_to_save = c(),
  cores = detect_cores(),
  pass_fit = F,
  do_check_only_on_detrimental = length(parse_formula(formula)) > 0,
  tol_rel_obj = 0.01,
  just_discovery = F,
  seed = sample(1:99999, size = 1),
  adj_prob_theshold_2 = NULL
)

identify_outliers(
  .data,
  formula = ~1,
  .sample,
  .transcript,
  .abundance,
  .significance,
  .do_check,
  percent_false_positive_genes = "1\%",
  how_many_negative_controls = 500,
  approximate_posterior_inference = T,
  approximate_posterior_analysis = NULL,
  draws_after_tail = 10,
  save_generated_quantities = F,
  additional_parameters_to_save = c(),
  cores = detect_cores(),
  pass_fit = F,
  do_check_only_on_detrimental = length(parse_formula(formula)) > 0,
  tol_rel_obj = 0.01,
  just_discovery = F,
  seed = sample(1:99999, size = 1),
  adj_prob_theshold_2 = NULL
)
}
\arguments{
\item{.data}{A tibble including a transcript name column | sample name column | read counts column | covariate columns | Pvaue column | a significance column}

\item{formula}{A formula. The sample formula used to perform the differential transcript abundance analysis}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.transcript}{A column name as symbol. The transcript identifier}

\item{.abundance}{A column name as symbol. The transcript abunace (read count)}

\item{.significance}{A column name as symbol. A column with the Pvalue, or other significanc measure (preferred Pvalue over false discovery rate)}

\item{.do_check}{A column name as symbol. A column with a booean indicating whether a transcript was identified as differentially abundant}

\item{percent_false_positive_genes}{A real}

\item{how_many_negative_controls}{An integer}

\item{approximate_posterior_inference}{A boolean}

\item{approximate_posterior_analysis}{A boolean}

\item{draws_after_tail}{An integer. How many draws should on average be after the tail, in a way to inform CI}

\item{save_generated_quantities}{A boolean}

\item{additional_parameters_to_save}{A character vector}

\item{cores}{An integer}

\item{pass_fit}{A boolean}

\item{do_check_only_on_detrimental}{A boolean}

\item{tol_rel_obj}{A real}

\item{just_discovery}{A boolean}

\item{seed}{an integer}

\item{adj_prob_theshold_2}{A boolean. Used for development and testing purposes}
}
\value{
A tibble with additional columns

A tibble with additional columns
}
\description{
This function calls the stan model.

This function calls the stan model.
}
