% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_results}
\alias{merge_results}
\title{merge_results}
\usage{
merge_results(
  res_discovery,
  res_test,
  formula,
  .transcript,
  .abundance,
  .sample,
  do_check_only_on_detrimental
)
}
\arguments{
\item{res_discovery}{A tibble}

\item{res_test}{A tibble}

\item{formula}{A formula}

\item{.transcript}{A column name}

\item{.abundance}{A column name}

\item{.sample}{A column name}

\item{do_check_only_on_detrimental}{A boolean}
}
\description{
merge_results
}
