% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcSeq.R
\name{produce_plots}
\alias{produce_plots}
\title{Produce generated quantities plots with marked uotliers}
\usage{
produce_plots(.x, value_column, sample_column, covariate)
}
\arguments{
\item{.x}{A tibble}

\item{value_column}{A symbol object}

\item{sample_column}{A symbol object}

\item{covariate}{A character string}
}
\value{
A ggplot
}
\description{
Produce generated quantities plots with marked uotliers
}
